uvuwe we ve oyno twe we ve obe mobe osas

+++ system/base/pam/actions.py
@@ -11,14 +11,14 @@ from inary.actionsapi import get
 
 def setup():
     inarytools.flags.add("-fPIC -D_GNU_SOURCE")
-    
+
     autotools.autoreconf("-fi")
-    autotools.configure("--libdir=/usr/lib \
+    autotools.configure("--libdir=/usr/lib{0} \
                          --enable-nls \
-                         --disable-audit \
-                         --enable-securedir=/lib/security \
-                         --enable-isadir=/lib/security")
-    
+                         --enable-audit \
+                         --enable-securedir=/lib{0}/security \
+                         --enable-isadir=/lib{0}/security".format("32" if get.buildTYPE()=="emul32" else ""))
+
     inarytools.dosed("libtool", " -shared ", " -Wl,-O1,--as-needed -shared ")
 
 def build():
@@ -36,6 +36,8 @@ def check():
 
 def install():
     autotools.rawInstall("DESTDIR=%s" % get.installDIR())
+    if get.buildTYPE()=="emul32":
+        return
 
     inarytools.removeDir("/usr/share/doc/Linux-PAM/")
